+++ src/client.rs
@@ -6,12 +6,17 @@ mod subscribe;
 
 pub use builder::*;
 
-use crate::{types::AccessToken, ArcSwapOptionExt};
+use crate::{ext::CookieJarExt, types::AccessToken, ArcSwapOptionExt};
 use arc_swap::ArcSwapOption;
+use cookie::{Cookie, CookieJar};
 use hyper::{
     client::HttpConnector, header::SET_COOKIE, http::HeaderValue, Body, Client, Response, Uri,
 };
-use std::sync::atomic::{AtomicUsize, Ordering};
+use std::{
+    borrow::Cow,
+    sync::atomic::{AtomicUsize, Ordering},
+};
+use tokio::sync::RwLock;
 
 /// A cometd Client.
 #[derive(Debug)]
@@ -25,7 +30,8 @@ pub struct CometdClient {
 
     id: AtomicUsize,
     pub(crate) access_token: ArcSwapOption<Box<dyn AccessToken>>,
-    pub(crate) cookie: ArcSwapOption<HeaderValue>,
+    pub(crate) cookies: RwLock<CookieJar>,
+    pub(crate) cookies_string_cache: ArcSwapOption<Box<str>>,
     client_id: ArcSwapOption<Box<str>>,
     pub(crate) http_client: Client<HttpConnector>,
 }
@@ -49,15 +55,56 @@ impl CometdClient {
         self.access_token.store_value(Box::new(access_token));
     }
 
+    /// Method for adding cookies.
+    ///
+    /// # Example
+    /// ```rust
+    /// # use cometd_client::{types::access_token::Basic, CometdClientBuilder};
+    /// # let client = CometdClientBuilder::new(&"http://[::1]:1025/".parse().unwrap()).build().unwrap();
+    ///
+    ///     client.add_cookies([("a", "1")]);
+    /// ```
+    #[inline(always)]
+    pub async fn add_cookies<N, V>(&self, cookies: impl IntoIterator<Item = (N, V)>)
+    where
+        N: Into<Cow<'static, str>>,
+        V: Into<Cow<'static, str>>,
+    {
+        let mut cookie_jar = self.cookies.write().await;
+        for (name, value) in cookies {
+            cookie_jar.add(Cookie::new(name, value));
+        }
+
+        self.cookies_string_cache
+            .store_value(cookie_jar.make_string());
+    }
+
     #[inline(always)]
     pub(crate) fn next_id(&self) -> String {
         self.id.fetch_add(1, Ordering::Relaxed).to_string()
     }
 
-    #[inline(always)]
-    pub(crate) async fn extract_and_store_cookie(&self, response: &mut Response<Body>) {
-        if let Some(cookie) = response.headers_mut().remove(SET_COOKIE) {
-            self.cookie.store_value(cookie);
+    #[inline]
+    pub(crate) async fn extract_and_store_cookie(&self, response: &Response<Body>) {
+        let mut redo_cache = false;
+
+        let mut cookies = self.cookies.write().await;
+        for cookie in response
+            .headers()
+            .get_all(SET_COOKIE)
+            .iter()
+            .map(HeaderValue::to_str)
+            .filter_map(Result::ok)
+            .map(str::to_string)
+            .map(Cookie::parse)
+            .filter_map(Result::ok)
+        {
+            cookies.add(cookie);
+            redo_cache = true;
+        }
+
+        if redo_cache {
+            self.cookies_string_cache.store_value(cookies.make_string());
         }
     }
 }