+++ src/common/request_builder.rs
@@ -1,23 +1,14 @@
 use crate::{consts::APPLICATION_JSON, types::AccessToken, CometdClient};
 use hyper::{
     header::{CONTENT_TYPE, COOKIE},
-    http::{request::Builder, HeaderValue},
+    http::request::Builder,
     Method, Request, Uri,
 };
 use std::sync::Arc;
 
 impl CometdClient {
-    #[inline(always)]
-    pub(crate) fn create_request_builder(&self, uri: &Uri) -> Builder {
-        self.create_request_builder_with_cookie(uri, self.cookie.load_full())
-    }
-
     #[inline]
-    pub(crate) fn create_request_builder_with_cookie(
-        &self,
-        uri: &Uri,
-        cookie: Option<Arc<HeaderValue>>,
-    ) -> Builder {
+    pub(crate) async fn create_request_builder(&self, uri: &Uri) -> Builder {
         let mut ret = Request::builder()
             .uri(uri)
             .method(Method::POST)
@@ -36,11 +27,8 @@ impl CometdClient {
         }
 
         // set cookies
-        if let Some(cookie) = cookie
-            .map(Arc::try_unwrap)
-            .map(|cookie| cookie.unwrap_or_else(|err| err.as_ref().clone()))
-        {
-            ret = ret.header(COOKIE, cookie);
+        if let Some(cookies) = self.cookies_string_cache.load().as_deref() {
+            ret = ret.header(COOKIE, cookies.as_ref());
         }
 
         ret