+++ CHANGELOG.md
@@ -11,6 +11,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 ### Add
 
 - Add access token setter for builder.
+- Add support for multiple cookies.
 
 ### Change